social_media_automation/requirements.txt
# Social Media API Libraries
instagrapi==2.1.2
facebook-sdk==3.1.0
requests==2.31.0
python-linkedin-v2==0.9.1

# Scheduling and Configuration
schedule==1.2.0
python-dotenv==1.0.0
pytz==2024.1

# Utilities
Pillow==10.2.0
